+++ client/tws_foundation_client/lib/src/services/business/trailers_classes_service.dart
@@ -1,5 +1,4 @@
 import 'package:csm_client/csm_client.dart';
-import 'package:tws_foundation_client/src/services/business/bases/trailers_classes_service_base.dart';
 import 'package:tws_foundation_client/tws_foundation_client.dart';
 
 final class TrailersClassesService extends TrailersClassesServiceBase {