+++ client/tws_foundation_client/test/definition/trailers_classes_service_test.dart
@@ -2,7 +2,6 @@ import 'dart:convert';
 
 import 'package:csm_client/csm_client.dart';
 import 'package:test/test.dart';
-import 'package:tws_foundation_client/src/services/business/bases/trailers_classes_service_base.dart';
 import 'package:tws_foundation_client/tws_foundation_client.dart';
 
 void main() {