+++ client/tws_foundation_client/test/integration/trailers_classes_service_test.dart
@@ -1,6 +1,5 @@
 import 'package:csm_client/csm_client.dart';
 import 'package:test/test.dart';
-import 'package:tws_foundation_client/src/services/business/bases/trailers_classes_service_base.dart';
 import 'package:tws_foundation_client/tws_foundation_client.dart';
 
 import '../integration_credentials.dart';